pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "react-agent"
version = "0.1.0"
description = "Local ReAct Agent with SMAK core, MCP client, and RAG-based memory"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "langchain-openai>=0.1",
    "mcp>=1.0",
]

[project.optional-dependencies]
smak = [
    "smak",
    "faiss-storage-lib",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
react-agent = "react_agent.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
